BaseAutofill: move to be it's own component (#3877)
* move base autofill and rename properly

* adding files

* adding change files

* renamed to autofill

* fixed rename

* added a few more comments
