+++ common/changes/@uifabric/experiments/joschect-move-baseautofill_2018-02-02-23-42.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "@uifabric/experiments",
+      "comment": "BasePicker: Use correct autofillnow",
+      "type": "patch"
+    }
+  ],
+  "packageName": "@uifabric/experiments",
+  "email": "joschect@microsoft.com"
+}
