+++ packages/office-ui-fabric-react/src/Autofill.ts
@@ -1,0 +1,1 @@
+export * from './components/Autofill/index';
