+++ packages/office-ui-fabric-react/src/components/Autofill/index.ts
@@ -1,0 +1,2 @@
+export * from './Autofill';
+export * from './Autofill.types';
